created model directory and moved model there
